Consolidate Group tests into one module
