--- tests/realtime/test_Group_extend.py
@@ -1,264 +1,0 @@
-import pytest
-import uqbar.strings
-
-import supriya.assets.synthdefs
-import supriya.realtime
-from supriya.osc import OscBundle, OscMessage
-
-
-@pytest.fixture(autouse=True)
-def shutdown_sync_servers(shutdown_scsynth):
-    pass
-
-
-@pytest.fixture
-def server(persistent_server):
-    persistent_server.reset()
-    persistent_server.add_synthdef(supriya.assets.synthdefs.default)
-    yield persistent_server
-
-
-def test_noop(server):
-    group = supriya.realtime.Group().allocate(server)
-    server_state = str(server.query())
-    assert server_state == uqbar.strings.normalize(
-        """
-        NODE TREE 0 group
-            1 group
-                1000 group
-        """
-    )
-    assert str(server.root_node) == server_state
-    with server.osc_protocol.capture() as transcript:
-        group.extend([])
-    assert [(_.label, _.message) for _ in transcript] == []
-    server_state = str(server.query())
-    assert server_state == uqbar.strings.normalize(
-        """
-        NODE TREE 0 group
-            1 group
-                1000 group
-        """
-    )
-    assert str(server.root_node) == server_state
-
-
-def test_allocate_nested(server):
-    group = supriya.realtime.Group()
-    synth_a = supriya.realtime.Synth(supriya.assets.synthdefs.test)
-    synth_b = supriya.realtime.Synth(supriya.assets.synthdefs.test, amplitude=0.0)
-    group.extend([synth_a, synth_b])
-    with server.osc_protocol.capture() as transcript:
-        group.allocate(server)
-    server_state = str(server.query())
-    assert server_state == uqbar.strings.normalize(
-        """
-        NODE TREE 0 group
-            1 group
-                1000 group
-                    1001 test
-                        amplitude: 1.0, frequency: 440.0
-                    1002 test
-                        amplitude: 0.0, frequency: 440.0
-        """
-    )
-    assert str(server.root_node) == server_state
-    bundle = OscBundle(
-        contents=(
-            OscMessage("/g_new", 1000, 0, 1),
-            OscMessage("/s_new", "test", 1001, 0, 1000),
-            OscMessage("/s_new", "test", 1002, 3, 1001, "amplitude", 0.0),
-        )
-    )
-    assert [(_.label, _.message) for _ in transcript] == [
-        (
-            "S",
-            OscMessage(
-                "/d_recv", bytearray(supriya.assets.synthdefs.test.compile()), bundle
-            ),
-        ),
-        ("R", OscMessage("/n_go", 1000, 1, -1, -1, 1, -1, -1)),
-        ("R", OscMessage("/n_go", 1001, 1000, -1, -1, 0)),
-        ("R", OscMessage("/n_go", 1002, 1000, 1001, -1, 0)),
-        ("R", OscMessage("/done", "/d_recv")),
-    ]
-
-
-def test_extend_unallocated(server):
-    group = supriya.realtime.Group()
-    group.allocate(server)
-    synth_a = supriya.realtime.Synth(supriya.assets.synthdefs.test)
-    synth_b = supriya.realtime.Synth(supriya.assets.synthdefs.test, amplitude=0.0)
-    with server.osc_protocol.capture() as transcript:
-        group.extend([synth_a, synth_b])
-    server_state = str(server.query())
-    assert server_state == uqbar.strings.normalize(
-        """
-        NODE TREE 0 group
-            1 group
-                1000 group
-                    1001 test
-                        amplitude: 1.0, frequency: 440.0
-                    1002 test
-                        amplitude: 0.0, frequency: 440.0
-        """
-    )
-    assert str(server.root_node) == server_state
-    bundle = OscBundle(
-        contents=(
-            OscMessage("/s_new", "test", 1001, 0, 1000),
-            OscMessage("/s_new", "test", 1002, 3, 1001, "amplitude", 0.0),
-        )
-    )
-    assert [(_.label, _.message) for _ in transcript] == [
-        ("S", OscMessage("/d_recv", supriya.assets.synthdefs.test.compile(), bundle)),
-        ("R", OscMessage("/n_go", 1001, 1000, -1, -1, 0)),
-        ("R", OscMessage("/n_go", 1002, 1000, 1001, -1, 0)),
-        ("R", OscMessage("/done", "/d_recv")),
-    ]
-
-
-def test_extend_allocate_nested_and_move(server):
-    synth = supriya.realtime.Synth().allocate(server)
-    synthdef = supriya.assets.synthdefs.test
-    group_a = supriya.realtime.Group().allocate(server)
-    group_b = supriya.realtime.Group(
-        [
-            supriya.realtime.Synth(synthdef=synthdef),
-            supriya.realtime.Group([supriya.realtime.Synth(synthdef=synthdef)]),
-        ]
-    )
-    server_state = str(server.query())
-    assert server_state == uqbar.strings.normalize(
-        """
-        NODE TREE 0 group
-            1 group
-                1001 group
-                1000 default
-                    out: 0.0, amplitude: 0.1, frequency: 440.0, gate: 1.0, pan: 0.5
-        """
-    )
-    assert str(server.root_node) == server_state
-    with server.osc_protocol.capture() as transcript:
-        group_a.extend([group_b, synth])
-    bundle = OscBundle(
-        contents=(
-            OscMessage("/g_new", 1002, 0, 1001),
-            OscMessage("/s_new", "test", 1003, 0, 1002),
-            OscMessage("/g_new", 1004, 3, 1003),
-            OscMessage("/s_new", "test", 1005, 0, 1004),
-            OscMessage("/n_after", 1000, 1002),
-        )
-    )
-    assert [(_.label, _.message) for _ in transcript] == [
-        ("S", OscMessage("/d_recv", supriya.assets.synthdefs.test.compile(), bundle)),
-        ("R", OscMessage("/n_go", 1002, 1001, -1, -1, 1, -1, -1)),
-        ("R", OscMessage("/n_go", 1003, 1002, -1, -1, 0)),
-        ("R", OscMessage("/n_go", 1004, 1002, 1003, -1, 1, -1, -1)),
-        ("R", OscMessage("/n_go", 1005, 1004, -1, -1, 0)),
-        ("R", OscMessage("/n_move", 1000, 1001, 1002, -1, 0)),
-        ("R", OscMessage("/done", "/d_recv")),
-    ]
-    server_state = str(server.query())
-    assert server_state == uqbar.strings.normalize(
-        """
-        NODE TREE 0 group
-            1 group
-                1001 group
-                    1002 group
-                        1003 test
-                            amplitude: 1.0, frequency: 440.0
-                        1004 group
-                            1005 test
-                                amplitude: 1.0, frequency: 440.0
-                    1000 default
-                        out: 0.0, amplitude: 0.1, frequency: 440.0, gate: 1.0, pan: 0.5
-        """
-    )
-    assert str(server.root_node) == server_state
-
-
-def test_x(server):
-    group_a = supriya.realtime.Group()
-    group_a.allocate(target_node=server)
-    group_b = supriya.realtime.Group()
-    group_b.allocate(target_node=server)
-    server_state = str(server.query())
-    assert server_state == uqbar.strings.normalize(
-        """
-        NODE TREE 0 group
-            1 group
-                1001 group
-                1000 group
-        """
-    )
-    synthdef = supriya.assets.synthdefs.test
-    assert synthdef not in server
-    synth_a = supriya.realtime.Synth(synthdef)
-    synth_b = supriya.realtime.Synth(synthdef)
-    synth_c = supriya.realtime.Synth(synthdef)
-    synth_d = supriya.realtime.Synth(synthdef)
-    assert not synth_a.is_allocated
-    assert not synth_b.is_allocated
-    assert not synth_c.is_allocated
-    assert not synth_d.is_allocated
-    synth_a.allocate(server)
-    server_state = str(server.query())
-    assert server_state == uqbar.strings.normalize(
-        """
-        NODE TREE 0 group
-            1 group
-                1002 test
-                    amplitude: 1.0, frequency: 440.0
-                1001 group
-                1000 group
-        """
-    )
-    group_a.extend([synth_a, synth_b])
-    server_state = str(server.query())
-    assert server_state == uqbar.strings.normalize(
-        """
-        NODE TREE 0 group
-            1 group
-                1001 group
-                1000 group
-                    1002 test
-                        amplitude: 1.0, frequency: 440.0
-                    1003 test
-                        amplitude: 1.0, frequency: 440.0
-        """
-    )
-    group_b.extend([synth_c])
-    server_state = str(server.query())
-    assert server_state == uqbar.strings.normalize(
-        """
-        NODE TREE 0 group
-            1 group
-                1001 group
-                    1004 test
-                        amplitude: 1.0, frequency: 440.0
-                1000 group
-                    1002 test
-                        amplitude: 1.0, frequency: 440.0
-                    1003 test
-                        amplitude: 1.0, frequency: 440.0
-        """
-    )
-    group_b.extend([synth_d, synth_b, synth_a])
-    server_state = str(server.query())
-    assert server_state == uqbar.strings.normalize(
-        """
-        NODE TREE 0 group
-            1 group
-                1001 group
-                    1004 test
-                        amplitude: 1.0, frequency: 440.0
-                    1005 test
-                        amplitude: 1.0, frequency: 440.0
-                    1003 test
-                        amplitude: 1.0, frequency: 440.0
-                    1002 test
-                        amplitude: 1.0, frequency: 440.0
-                1000 group
-        """
-    )
